Refactoring: move experimental imports checking to Checking.scala
